CTPA-291 more good stuff

+++ caseframesvc-api/src/main/java/uk/gov/ons/ctp/response/caseframe/representation/CaseDTO.java
@@ -21,7 +21,7 @@ public class CaseDTO {
 
   private Integer caseId;
 
-  private Integer uprn;
+  private Long uprn;
 
   private CaseState state;
